fix(LIVE-34547): fix duplicated
