+++ apps/ledger-live-mobile/src/mvvm/features/Exchange/index.ts
@@ -1,0 +1,5 @@
+export {
+  getAccountsForCurrencies,
+  resolveCurrencyIds,
+  type AccountWithParent,
+} from "./utils/currencyAccounts";
